Fix Wko GPS parsing on 64bit
This bug recently fixed in v3 is also present in the
v2 code.

Fixes #214.

+++ src/WkoRideFile.cpp
@@ -363,7 +363,7 @@ WKO_UCHAR *WkoParseRawData(WKO_UCHAR *fb, RideFile *rideFile, QStringList &error
                         break;
                     case 'G' : /* two longs */
                         {
-                            signed long llat, llon;
+                            int32_t llat, llon;
                             char slat[20], slon[20];
 
                             // stored 2s complement
